seccomp: refactor package structure
Refactor the general seccomp package structure, in order
to restrict public access to the seccompiler internals.

Previously, the library crate was solely in the lib.rs file,
and it was impossible to restrict the public visibility of the
unneeded structures, since the binary crate from the same folder
needed access to those data structures.

Now, the items that need to be exported to the public library are
still in the lib.rs file. The other internal functionality and
data types were moved to backend.rs.

Data types that are common between the lib and bin crates are now
in the common.rs file.

This achieves a minimal lib.rs file, exposing the desired
helper functions and nothing more.

To summarise the folder structure:

- lib.rs -> public lib interface (with helper functions used by FC)
- common.rs -> data types used by both the bin and lib crates.
- seccompiler.rs -> main() binary, for compiling JSON into BPF.
- backend.rs -> used by the binary for bpf generation from IR.
- compiler.rs -> used by the binary for driving the enire compilation
- syscall_table module -> used by the binary for mapping syscall
        names to the arch-specific number.

Also restricted the pub to pub(crate) or private, where applicable.

Also updated the respective cargo-docs.
